Moved around some style files
Better sorting
