+++ src/styles/components/ui/button.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import generalStyleInteractions from './generalStylesInteractions';
+import generalStyleInteractions from './multi-use/generalStylesInteractions';
 
 export const Button = styled.button`
   ${generalStyleInteractions}
