+++ src/styles/layouts/SearchMediaLayout.tsx
